Completed assignment

+++ Binomial.java
@@ -21,7 +21,7 @@ return binomial1(n - 1, k) + binomial1(n - 1, k - 1);
 	 
 	
 	// Computes the Binomial function, efficiently
-	public static long binomial(int n, int k) {
+	public static int binomial(int n, int k) {
 		int [][] memo = new int [n + 1][k + 1];
 		 for(int i = 0; i < memo.length; i++) {
 			for(int j = 0; j < memo[0].length; j++) {
@@ -32,7 +32,7 @@ return binomial1(n - 1, k) + binomial1(n - 1, k - 1);
 	}
 
 
-	private static long binomial(int n, int k, long[][] memo) {
+	private static int binomial(int n, int k, int[][] memo) {
 		if (memo[n][k] != -1) {
 			return memo[n][k];
 		}
